Fixes for comments
