+++ JDI.Light/JDI.Light/JDI.Light.csproj
@@ -74,6 +74,7 @@
     <Compile Include="Elements\Composite\Alert.cs" />
     <Compile Include="Elements\Composite\Menu.cs" />
     <Compile Include="Elements\Composite\MultiDropdown.cs" />
+    <Compile Include="Elements\Composite\MultiDropdownElement.cs" />
     <Compile Include="Elements\Composite\Table.cs" />
     <Compile Include="Elements\Composite\TextList.cs" />
     <Compile Include="Enums\CheckPageType.cs" />
